+++ src/SymbolEditor/SymbolEditor.MSIX/Images/LockScreenLogo.scale-200.png
