+++ src/SymbolEditor/SymbolEditor.MSIX/Images/SplashScreen.scale-200.png
