+++ src/SymbolEditor/SymbolEditor.MSIX/Images/Square150x150Logo.scale-200.png
