+++ src/SymbolEditor/SymbolEditor.MSIX/Images/Square44x44Logo.scale-200.png
