+++ src/SymbolEditor/SymbolEditor.MSIX/Images/Square44x44Logo.targetsize-24_altform-unplated.png
