+++ src/SymbolEditor/SymbolEditor.MSIX/Images/StoreLogo.png
