+++ src/SymbolEditor/SymbolEditor.MSIX/Images/Wide310x150Logo.scale-200.png
