+++ src/SymbolEditor/SymbolEditor.MSIX/Package.appxmanifest
