+++ src/SymbolEditor/SymbolEditor.MSIX/SymbolEditor.MSIX.wapproj
@@ -55,7 +55,7 @@
     <TargetPlatformMinVersion>10.0.16299.0</TargetPlatformMinVersion>
     <DefaultLanguage>en-US</DefaultLanguage>
     <AppxPackageSigningEnabled>false</AppxPackageSigningEnabled>
-    <EntryPointProjectUniqueName>..\MobileStylePicker\MobileStylePicker.csproj</EntryPointProjectUniqueName>
+    <EntryPointProjectUniqueName>..\MobileStylePicker\SymbolEditorApp.csproj</EntryPointProjectUniqueName>
   </PropertyGroup>
   <ItemGroup>
     <AppxManifest Include="Package.appxmanifest">
@@ -72,7 +72,7 @@
     <Content Include="Images\Wide310x150Logo.scale-200.png" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\MobileStylePicker\MobileStylePicker.csproj">
+    <ProjectReference Include="..\MobileStylePicker\SymbolEditorApp.csproj">
       <SkipGetTargetFrameworkProperties>True</SkipGetTargetFrameworkProperties>
     </ProjectReference>
   </ItemGroup>
