+++ src/SymbolEditor/SymbolEditorApp/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="MobileStylePicker.App"
+﻿<Application x:Class="SymbolEditorApp.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:MobileStylePicker"
+             xmlns:local="clr-namespace:SymbolEditorApp"
              Startup="Application_Startup"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
