+++ src/SymbolEditor/SymbolEditorApp/App.xaml.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 
-namespace MobileStylePicker
+namespace SymbolEditorApp
 {
     /// <summary>
     /// Interaction logic for App.xaml
