+++ src/SymbolEditor/SymbolEditorApp/Controls/SymbolEditor.xaml
@@ -1,10 +1,10 @@
-﻿<UserControl x:Class="MobileStylePicker.SymbolEditor"
+﻿<UserControl x:Class="SymbolEditorApp.Controls.SymbolEditor"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:mah="http://metro.mahapps.com/winfx/xaml/controls"
-             xmlns:local="clr-namespace:MobileStylePicker"
+             xmlns:local="clr-namespace:SymbolEditorApp.Controls"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
              mc:Ignorable="d"
              d:DesignHeight="450" d:DesignWidth="800">
