+++ src/SymbolEditor/SymbolEditorApp/Controls/SymbolEditor.xaml.cs
@@ -13,7 +13,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace MobileStylePicker
+namespace SymbolEditorApp.Controls
 {
     /// <summary>
     /// Interaction logic for SymbolEditor.xaml
