+++ src/SymbolEditor/SymbolEditorApp/Controls/SymbolPicker.xaml
@@ -1,9 +1,9 @@
-﻿<mah:MetroWindow x:Class="MobileStylePicker.SymbolPicker"
+﻿<mah:MetroWindow x:Class="SymbolEditorApp.Controls.SymbolPicker"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MobileStylePicker"
+             xmlns:local="clr-namespace:SymbolEditorApp.Controls"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
              xmlns:mah="http://metro.mahapps.com/winfx/xaml/controls"
              mc:Ignorable="d" ResizeMode="NoResize"
