+++ src/SymbolEditor/SymbolEditorApp/Controls/SymbolViewer.xaml
@@ -1,9 +1,8 @@
-﻿<UserControl x:Class="MobileStylePicker.SymbolViewer"
+﻿<UserControl x:Class="SymbolEditorApp.Controls.SymbolViewer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:MobileStylePicker"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
              mc:Ignorable="d" 
              Width="300" Height="320"
