+++ src/SymbolEditor/SymbolEditorApp/Controls/SymbolViewer.xaml.cs
@@ -15,7 +15,7 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace MobileStylePicker
+namespace SymbolEditorApp.Controls
 {
     /// <summary>
     /// Interaction logic for SymbolViewer.xaml
