+++ src/SymbolEditor/SymbolEditorApp/MainWindow.xaml
@@ -1,10 +1,10 @@
-﻿<mah:MetroWindow x:Class="MobileStylePicker.MainWindow"
+﻿<mah:MetroWindow x:Class="SymbolEditorApp.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013"
-        xmlns:local="clr-namespace:MobileStylePicker"
+        xmlns:local="clr-namespace:SymbolEditorApp"
         xmlns:mah="http://metro.mahapps.com/winfx/xaml/controls"
         mc:Ignorable="d"
         Title="Symbol Style Picker" Height="768" Width="1024">
@@ -19,6 +19,6 @@
         </Grid.ColumnDefinitions>
         <esri:MapView Map="{Binding Map}" Grid.Column="2" x:Name="mapView" />
         <GridSplitter Grid.Column="1" ResizeDirection="Auto" />
-        <esri:TableOfContents GeoView="{Binding ElementName=mapView}" TableOfContentContextMenuOpening="TableOfContents_TableOfContentContextMenuOpening"  />
+        <esri:TableOfContents GeoView="{Binding ElementName=mapView}" Margin="5" TableOfContentContextMenuOpening="TableOfContents_TableOfContentContextMenuOpening"  />
     </Grid>
 </mah:MetroWindow>
