+++ src/SymbolEditor/SymbolEditorApp/MainWindow.xaml.cs
@@ -1,21 +1,11 @@
 ﻿using Esri.ArcGISRuntime.Mapping;
 using Esri.ArcGISRuntime.Symbology;
+using SymbolEditorApp.Controls;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
-namespace MobileStylePicker
+namespace SymbolEditorApp
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
