+++ src/SymbolEditor/SymbolEditorApp/Resources/ArcGISRuntime2D_Pro25.stylx
