+++ src/SymbolEditor/SymbolEditorApp/SymbolEditorApp.csproj
@@ -4,6 +4,7 @@
     <OutputType>WinExe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <UseWPF>true</UseWPF>
+    <AssemblyName>SymbolEditorApp</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Content Include="ArcGISRuntime2D_Pro25.stylx">
+    <Content Include="Resources\ArcGISRuntime2D_Pro25.stylx">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
